Hardcode cycles into every instruction

+++ src/emulator/bus.rs
@@ -169,7 +169,7 @@ impl Bus {
         self.interrupts.master = false;
     }
 
-    pub fn run_system(&mut self, cycles: u16) {
+    pub fn run_system(&mut self, cycles: u8) {
         self.gpu.step(cycles, &mut self.interrupts);
         self.timer.step(cycles, &mut self.interrupts);
         //self.sound.step
