+++ src/emulator/mod.rs
@@ -115,7 +115,7 @@ impl Gameboy {
 
 
     //execute instruction pointed by PC, increment it as needed, return number of cycles it took and if an IO write was made
-    fn cpu_inst(&mut self, debug_flag: bool) -> u16 {
+    fn cpu_inst(&mut self, debug_flag: bool) -> u8 {
         self.cpu.interrupts(&mut self.bus);
         
         if !self.bus.halt_cpu {
